Add the 'nixtures' directory, for _not_ value fixtures
I'd normally make this in 'fixtures/invalid', but fixtures seems to be
well-established as containing only valid/parseable files already, and I
don't want to suggest such a broad change. But 'nix' is negative, so I
can create another fixtures directory that means 'bad fixtures' and is
also a pun!